% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{beta_best_dens}
\alias{beta_best_dens}
\title{Evaluate joint density of maximum of Beta RVs}
\usage{
beta_best_dens(x, a, b, minimum = FALSE)
}
\arguments{
\item{x}{Value to evaluate density at}

\item{a}{Vector of shape1 parameters}

\item{b}{Vector of shape1=2 parameters}

\item{minimum}{Use minimum instead of maximum.}
}
\description{
Evaluate joint density of maximum of Beta RVs
}
