% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_brar_design.R
\name{beta_brar_trial}
\alias{beta_brar_trial}
\title{Run Bayesian response-adaptive randomised trial}
\usage{
beta_brar_trial(
  n_seq,
  dat,
  rand_quant = "best",
  active_quant = "best",
  control = "fix",
  approx = FALSE,
  perm_drop = FALSE,
  mwu = FALSE,
  minimum = TRUE,
  bpar = 0.5,
  hpar1 = 2.5,
  hpar2 = 3,
  h = function(i) hpar1 * (i/length(n_seq))^hpar2,
  fpar1 = 0.1,
  fpar2 = 1.5,
  f = function(i) fpar1 * (i/length(n_seq))^fpar2,
  ...
)
}
\arguments{
\item{n_seq}{Sequence of interim analyses}

\item{dat}{Potential outcomes data from \code{gen_potential_outcomes}}

\item{rand_quant}{Quantity to decide allocations, \code{best} or \code{superior}}

\item{active_quant}{Quantity to decide active arms, \code{best} or \code{superior}}

\item{control}{How to allocate to control, \code{fix}, \code{constant}, \code{match}.}

\item{approx}{Use approximate rather than numerical methods.}

\item{perm_drop}{Permanently drop any arm which is not \code{active}.}

\item{mwu}{Use mass-weighted urn randomisation}

\item{bpar}{Parameter for control = \code{match}.}

\item{hpar1}{Parameter to control amount of RAR}

\item{hpar2}{Parameter to control scaling of RAR}

\item{h}{Function for RAR}

\item{fpar1}{Parameter to control lack of effect boundary (futility)}

\item{fpar2}{Parameter to control lack of effect boundary (futility)}

\item{f}{Function for utility boundary}

\item{...}{Other function arguments}
}
\description{
Run Bayesian response-adaptive randomised trial
}
