% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{beta_prob_best}
\alias{beta_prob_best}
\title{Probability that each element of a collection of Beta RV is the maximum/minimum}
\usage{
beta_prob_best(a, b, minimum = FALSE, approx = FALSE, ...)
}
\arguments{
\item{a}{Vector of shape1 parameters}

\item{b}{Vector of shape2 parameters}

\item{minimum}{Use minimum instead of maximum}

\item{approx}{Use approximation instead of integration}

\item{...}{Other arguments to \code{pracma::quadgk}}
}
\value{
A vector of \code{length(a)} giving \verb{Pr(a[i] is max)}
}
\description{
Probability that each element of a collection of Beta RV is the maximum/minimum
}
