% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{beta_prob_best_numeric}
\alias{beta_prob_best_numeric}
\title{Probability that each element of a collection of Beta RV is the maximum/minimum}
\usage{
beta_prob_best_numeric(a, b, minimum = FALSE, ...)
}
\arguments{
\item{a}{Vector of shape1 parameters}

\item{b}{Vector of shape2 parameters}

\item{minimum}{Use minimum instead of maximum}

\item{...}{Other arguments to \code{pracma::quadgk}}
}
\value{
A vector of \code{length(a)} giving \verb{Pr(a[i] is max)}
}
\description{
Probability that each element of a collection of Beta RV is the maximum/minimum
}
\examples{
beta_prob_best_numeric(1 + seq(1, 9, 2), 1 + rep(10, 5))
}
