% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{beta_prob_supr}
\alias{beta_prob_supr}
\title{Probability that each arm superior to reference.}
\usage{
beta_prob_supr(a, b, reverse = FALSE, approx = FALSE, ...)
}
\arguments{
\item{a}{First value is reference}

\item{b}{First value is reference}

\item{reverse}{Reverse direction of superiority}

\item{approx}{Use normal approximation instead of numerical integration}

\item{...}{Other arguments to \code{pracma::quadgk}}
}
\description{
Calculates event that one Beta RV larger (or smaller) to another Beta RV by normal approximation.
}
\examples{
beta_prob_supr(c(5, 10, 15), c(15, 10, 5))
beta_prob_supr(c(5, 10, 15), c(15, 10, 5), approx = TRUE)
}
