% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{beta_prob_supr_numeric}
\alias{beta_prob_supr_numeric}
\title{Probability that each arm superior to reference.}
\usage{
beta_prob_supr_numeric(a, b, reverse = FALSE, ...)
}
\arguments{
\item{a}{First value is reference}

\item{b}{First value is reference}

\item{reverse}{Reverse direction of superiority}

\item{...}{Other arguments to \code{pracma::quadgk}}
}
\description{
Calculates event that one Beta RV larger (or smaller if reverse) to another Beta RV by integration.
}
\examples{
beta_prob_supr_numeric(c(5, 10, 15), c(15, 10, 5))
beta_prob_supr_numeric(c(5, 10, 15), c(15, 10, 5), reverse = TRUE)
}
