% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{beta_var}
\alias{beta_var}
\title{Variance of a beta random variable with parameters a and b}
\usage{
beta_var(a, b)
}
\arguments{
\item{a}{Par 1}

\item{b}{Par 2}
}
\description{
Variance of a beta random variable with parameters a and b
}
\examples{
beta_var(5, 5)
}
