% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{brar_all}
\alias{brar_all}
\title{BRAR on all arms.}
\usage{
brar_all(quantity, active, h)
}
\arguments{
\item{quantity}{The quantity to be used for allocation ratios.}

\item{active}{Flag indicating active arms.}

\item{h}{Scaling factor value}
}
\value{
Allocation probabilities to each arm.
}
\description{
BRAR on all arms.
}
\examples{
brar_all(runif(5), sample(c(TRUE, FALSE), 5, rep = TRUE), 1)
}
