% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{const_ctrl_brar}
\alias{const_ctrl_brar}
\title{BRAR with constant proportion to control}
\usage{
const_ctrl_brar(quantity, active, h)
}
\arguments{
\item{quantity}{The quantity to be used for allocation ratios.}

\item{active}{Flag indicating active arms.}

\item{h}{Scaling factor value}
}
\value{
Allocation probabilities to each arm.
}
\description{
BRAR with constant proportion to control
}
\examples{
const_ctrl_brar(runif(5), sample(c(TRUE, FALSE), 5, rep = TRUE), 1)
}
