% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{ent_best_arm_approx}
\alias{ent_best_arm_approx}
\title{Entropy of the response density of the unknown maximum}
\usage{
ent_best_arm_approx(a, b, minimum = FALSE, nn = 100)
}
\arguments{
\item{a}{Vector of shape1 parameters}

\item{b}{Vector of shape2 parameters}

\item{minimum}{If true use minimum instead of maximum.}

\item{nn}{Grid points.}
}
\description{
Entropy of the response density of the unknown maximum
}
