% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{ent_best_arm_numeric}
\alias{ent_best_arm_numeric}
\title{Entropy of the response density of the unknown maximum}
\usage{
ent_best_arm_numeric(a, b, minimum = FALSE, ...)
}
\arguments{
\item{a}{Vector of shape1 parameters}

\item{b}{Vector of shape2 parameters}

\item{minimum}{If true use minimum instead of maximum.}

\item{...}{Other arguments to \code{quadgk}.}
}
\description{
Entropy of the response density of the unknown maximum
}
