% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{findfirst}
\alias{findfirst}
\title{Find first element satisfying condition}
\usage{
findfirst(x, v = NA)
}
\arguments{
\item{x}{A logical vector}

\item{v}{A default value if none satisfy condition}
}
\value{
Index of first to satisfy condition
}
\description{
Find first element satisfying condition
}
