% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{fix_ctrl_brar}
\alias{fix_ctrl_brar}
\title{BRAR with fixed ratio to control}
\usage{
fix_ctrl_brar(quantity, active, h)
}
\arguments{
\item{quantity}{The quantity for allocation ratios}

\item{active}{Flag for active arms}

\item{h}{Scaling factor}
}
\value{
A vector of allocation probabilities
}
\description{
BRAR with fixed ratio to control
}
