% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{fix_ctrl_brar2}
\alias{fix_ctrl_brar2}
\title{BRAR with fixed ratio to control}
\usage{
fix_ctrl_brar2(quantity, active, h, fixq)
}
\arguments{
\item{quantity}{The quantity for allocation ratios}

\item{active}{Flag for active arms}

\item{h}{Scaling factor}

\item{fixq}{The quantity to fix control at}
}
\description{
BRAR with fixed ratio to control
}
