% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{gen_potential_outcomes}
\alias{gen_potential_outcomes}
\title{Generate potential binary outcomes}
\usage{
gen_potential_outcomes(
  n,
  p,
  accrual_rate = function(n) (1:n)/2,
  delay = 0,
  ...
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Vector of probabilities of outcome}

\item{accrual_rate}{A function target integer argument which
generates that number of arrival times.}

\item{delay}{The time until endpoint is observed.}

\item{...}{Optional additional arguments to the accrual_rate function.}
}
\value{
A matrix of potential outcomes for each arm
}
\description{
Generate potential binary outcomes
}
