% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{mass_weighted_urn_design}
\alias{mass_weighted_urn_design}
\title{Mass-weighted urn randomisation}
\usage{
mass_weighted_urn_design(target_alloc, sample_size, alpha = 4)
}
\arguments{
\item{target_alloc}{The target allocation ratios}

\item{sample_size}{The number of allocations to generate}

\item{alpha}{Parameter to control imbalance between arms}
}
\value{
A list detailing the mass-weighted-urn process.
}
\description{
Mass-weighted urn randomisation
}
