% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{match_ctrl_brar}
\alias{match_ctrl_brar}
\title{BRAR with matching control on biggest sample size (Trippa)}
\usage{
match_ctrl_brar(quantity, active, n, h, bpar = 1)
}
\arguments{
\item{quantity}{The quantity for allocation ratios}

\item{active}{Flag for active arms}

\item{n}{Sample size in each arm}

\item{h}{Scaling factor}

\item{bpar}{Control matching parameter}
}
\description{
BRAR with matching control on biggest sample size (Trippa)
}
