% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{normalise}
\alias{normalise}
\title{Normalise vector to sum to 1}
\usage{
normalise(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The vector \code{x} normalised to sum to 1.
}
\description{
Normalise vector to sum to 1
}
