% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{prob_best}
\alias{prob_best}
\title{Probability each column is maximum (or minimum)}
\usage{
prob_best(mat, minimum = F)
}
\arguments{
\item{mat}{A matrix of draws.}

\item{minimum}{If \code{TRUE} determine best based on minimum instead of maximum value.}
}
\description{
Probability each column is maximum (or minimum)
}
