% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{psolve}
\alias{psolve}
\title{Basic Moore-Penrose pseudo-inverse}
\usage{
psolve(X)
}
\arguments{
\item{X}{A matrix}
}
\value{
Pseudo-inverse of X
}
\description{
Basic Moore-Penrose pseudo-inverse
}
