% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{utility_var}
\alias{utility_var}
\title{Calculate change in utility using variance}
\usage{
utility_var(n, y, a = 1, b = 1)
}
\arguments{
\item{n}{Sample size in each arm}

\item{y}{Response in each arm}

\item{a}{Prior shape}

\item{b}{Prior shape}
}
\description{
Calculate change in utility using variance
}
